requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.12.4
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
python-dotenv==1.0.1
drf-spectacular==0.27.1